+++ src/components/create-poll.tsx
@@ -6,6 +6,7 @@ import posthog from "posthog-js";
 import React from "react";
 import { useSessionStorage } from "react-use";
 
+import StandardLayout from "./layouts/standard-layout";
 import { encodeDateOption } from "../utils/date-time-utils";
 import { trpc } from "../utils/trpc";
 import { Button } from "./button";
@@ -18,7 +19,6 @@ import {
   UserDetailsData,
   UserDetailsForm,
 } from "./forms";
-import StandardLayout from "./standard-layout";
 import Steps from "./steps";
 import { useUser } from "./user-provider";
