+++ src/pages/profile.tsx
@@ -3,8 +3,8 @@ import { NextPage } from "next";
 import { withSessionSsr } from "@/utils/auth";
 
 import { Profile } from "../components/profile";
-import StandardLayout from "../components/standard-layout";
 import { withSession } from "../components/user-provider";
+import StandardLayout from "../components/layouts/standard-layout";
 import { withPageTranslations } from "../utils/with-page-translations";
 
 const Page: NextPage = () => {
